code modified for long integers

+++ Binary_Search/src/SQRT_69.java
@@ -17,11 +17,11 @@ public class SQRT_69 {
 
         while(start <= end)
         {   int mid = start + (end - start) / 2;
-            if(mid*mid == x)
+            if((long)mid*mid == x)
             {
                 return mid;
             }
-            else if(mid*mid < x)
+            else if((long)mid*mid < x)
             {
                 ans = mid;
                 start = mid+1;
